src/experience/requirements.txt
django-cors-headers>=2.4.0 
requests==2.23.0